Make file positions make more sense
